Move startup warnings into separate module
